backend/requirements/base.txt
argon2-cffi==21.3.0
crispy-bootstrap5==0.7
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-guardian==2.4.0
django-pwa==1.1.0
oauthlib==3.2.2
Pillow==9.3.0
python-decouple==3.6
